user-story-tests/25-inherit-imported-css-file.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Inherit @imported CSS file</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        @import url(assets/resets.css) layer(resets);
    </style>
    <button>Button outside a shadow tree</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer inherit.resets.as.outerresets, outerresets;
            </style>
            <button>
                Button inside a shadow tree (thick dashed red) (from inherited
                resets.css file)
            </button>
        </template>
    </button-group>

    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = ShadowLayers.inheritFromShadowStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");
            }, "Inherit @imported CSS file");

        }
    </script>

    <div id="log"></div>

</body>

</html>
